Updated the diagrams
