+++ ORM/src/main/java/com/salanb/orm/session/Session.java
@@ -1,4 +1,6 @@
 package com.salanb.orm.session;
 
-public class Session {
+public interface Session {
+
+    public void close();
 }
