+++ ORM/src/main/java/com/salanb/orm/session/SessionFactory.java
@@ -2,15 +2,7 @@ package com.salanb.orm.session;
 
 import com.salanb.orm.configuration.Configuration;
 
-public class SessionFactory {
+public interface SessionFactory {
 
-    public SessionFactory(Configuration config) {
-
-        /// TODO: Strip out the white space and initialize the Session Factory
-
-    }
-
-    public Session getSession() {
-        return new Session();
-    }
+    public SessionImplementation getSession();
 }
