+++ ORM/src/main/java/com/salanb/orm/session/SessionImplementation.java
@@ -1,0 +1,10 @@
+package com.salanb.orm.session;
+
+public class SessionImplementation implements Session{
+
+
+    @Override
+    public void close() {
+
+    }
+}
