+++ ORM/src/main/java/com/salanb/orm/utillities/Identifier.java
@@ -1,0 +1,6 @@
+package com.salanb.orm.utillities;
+
+import java.util.LinkedList;
+
+public class Identifier extends LinkedList<Object> implements IdentifierInterface {
+}
