+++ ORM/src/main/java/com/salanb/orm/utillities/IdentifierInterface.java
@@ -1,0 +1,4 @@
+package com.salanb.orm.utillities;
+
+public interface IdentifierInterface {
+}
